+++ src/type-data-frame.c
@@ -2,6 +2,8 @@
 #include "type-data-frame.h"
 #include "utils.h"
 
+static SEXP new_compact_rownames(R_len_t n);
+
 
 // [[ include("type-data-frame.h") ]]
 bool is_data_frame(SEXP x) {
@@ -44,7 +46,8 @@ SEXP new_data_frame(SEXP x, R_len_t n) {
 }
 
 static R_len_t df_size_from_list(SEXP x, SEXP n);
-static void poke_data_frame_class(SEXP x, SEXP cls);
+static R_len_t df_size_from_n(SEXP n);
+static SEXP c_data_frame_class(SEXP cls);
 
 // [[ register() ]]
 SEXP vctrs_new_data_frame(SEXP args) {
@@ -55,23 +58,85 @@ SEXP vctrs_new_data_frame(SEXP args) {
   SEXP cls = CAR(args); args = CDR(args);
   SEXP attrib = args;
 
+  PROTECT_INDEX pi;
+  PROTECT_WITH_INDEX(attrib, &pi);
+
   if (TYPEOF(x) != VECSXP) {
     Rf_errorcall(R_NilValue, "`x` must be a list");
   }
 
+  bool has_names = false;
+  bool has_rownames = false;
   R_len_t size = df_size_from_list(x, n);
 
-  if (attrib != R_NilValue) {
-    x = r_maybe_duplicate(x);
-    SET_ATTRIB(x, attrib);
-  }
-  PROTECT(x);
+  SEXP out = PROTECT(r_maybe_duplicate(x));
 
-  SEXP out = PROTECT(new_data_frame(x, size));
+  for (SEXP node = attrib; node != R_NilValue; node = CDR(node)) {
+    SEXP tag = TAG(node);
 
-  if (cls != R_NilValue) {
-    poke_data_frame_class(out, cls);
+    // We might add dynamic dots later on
+    if (tag == R_ClassSymbol) {
+      Rf_error("Internal error in `new_data_frame()`: Can't supply `class` in `...`.");
+    }
+
+    if (tag == R_NamesSymbol) {
+      has_names = true;
+      continue;
+    }
+
+    if (tag == R_RowNamesSymbol) {
+      // "row.names" is checked for consistency with n (if provided)
+      if (size != rownames_size(CAR(node))) {
+        Rf_errorcall(R_NilValue, "`n` and `row.names` must be consistent.");
+      }
+
+      has_rownames = true;
+      continue;
+    }
   }
+
+  // Take names from `x` if `attrib` doesn't have any
+  if (!has_names) {
+    SEXP nms = vctrs_shared_empty_chr;
+    if (Rf_length(out)) {
+      nms = r_names(out);
+    }
+    PROTECT(nms);
+
+    if (nms != R_NilValue) {
+      attrib = Rf_cons(nms, attrib);
+      SET_TAG(attrib, R_NamesSymbol);
+      REPROTECT(attrib, pi);
+    }
+
+    UNPROTECT(1);
+  }
+
+  if (!has_rownames) {
+    SEXP rn = PROTECT(new_compact_rownames(size));
+    attrib = Rf_cons(rn, attrib);
+    SET_TAG(attrib, R_RowNamesSymbol);
+
+    UNPROTECT(1);
+    REPROTECT(attrib, pi);
+  }
+
+  if (cls == R_NilValue) {
+    cls = classes_data_frame;
+  } else {
+    cls = c_data_frame_class(cls);
+  }
+  PROTECT(cls);
+
+  attrib = Rf_cons(cls, attrib);
+  SET_TAG(attrib, R_ClassSymbol);
+
+  UNPROTECT(1);
+  REPROTECT(attrib, pi);
+
+
+  SET_ATTRIB(out, attrib);
+  SET_OBJECT(out, 1);
 
   UNPROTECT(2);
   return out;
@@ -81,7 +146,10 @@ static R_len_t df_size_from_list(SEXP x, SEXP n) {
   if (n == R_NilValue) {
     return df_raw_size_from_list(x);
   }
+  return df_size_from_n(n);
+}
 
+static R_len_t df_size_from_n(SEXP n) {
   if (TYPEOF(n) != INTSXP || Rf_length(n) != 1) {
     Rf_errorcall(R_NilValue, "`n` must be an integer of size 1");
   }
@@ -89,31 +157,27 @@ static R_len_t df_size_from_list(SEXP x, SEXP n) {
   return r_int_get(n, 0);
 }
 
-static void poke_data_frame_class(SEXP x, SEXP cls) {
-  if (cls == R_NilValue) {
-    return;
-  }
+static SEXP c_data_frame_class(SEXP cls) {
   if (TYPEOF(cls) != STRSXP) {
     Rf_errorcall(R_NilValue, "`class` must be NULL or a character vector");
   }
   if (Rf_length(cls) == 0) {
-    return;
+    return classes_data_frame;
   }
 
   SEXP args = PROTECT(Rf_allocVector(VECSXP, 2));
   SET_VECTOR_ELT(args, 0, cls);
   SET_VECTOR_ELT(args, 1, classes_data_frame);
 
-  cls = PROTECT(vec_c(
+  SEXP out = vec_c(
     args,
     vctrs_shared_empty_chr,
     R_NilValue,
     NULL
-  ));
+  );
 
-  Rf_setAttrib(x, R_ClassSymbol, cls);
-
-  UNPROTECT(2);
+  UNPROTECT(1);
+  return out;
 }
 
 // [[ include("type-data-frame.h") ]]
@@ -133,13 +197,24 @@ enum rownames_type rownames_type(SEXP x) {
   }
 }
 
-// [[ include("type-data-frame.h") ]]
-R_len_t compact_rownames_length(SEXP x) {
+static R_len_t compact_rownames_length(SEXP x) {
   return abs(INTEGER(x)[1]);
 }
 
+// [[ include("type-data-frame.h") ]]
+R_len_t rownames_size(SEXP rn) {
+  switch (rownames_type(rn)) {
+  case ROWNAMES_IDENTIFIERS:
+  case ROWNAMES_AUTOMATIC:
+    return Rf_length(rn);
+  case ROWNAMES_AUTOMATIC_COMPACT:
+    return compact_rownames_length(rn);
+  }
+
+  never_reached("rownames_size");
+}
+
 static void init_bare_data_frame(SEXP x, R_len_t n);
-static SEXP new_compact_rownames(R_len_t n);
 
 // [[ include("type-data-frame.h") ]]
 void init_data_frame(SEXP x, R_len_t n) {