+++ src/type-data-frame.h
@@ -9,7 +9,6 @@ void init_compact_rownames(SEXP x, R_len_t n);
 SEXP df_rownames(SEXP x);
 
 bool is_native_df(SEXP x);
-R_len_t compact_rownames_length(SEXP x);
 SEXP df_container_type(SEXP x);
 SEXP df_poke(SEXP x, R_len_t i, SEXP value);
 SEXP df_poke_at(SEXP x, SEXP name, SEXP value);
@@ -23,6 +22,7 @@ enum rownames_type {
   ROWNAMES_IDENTIFIERS
 };
 enum rownames_type rownames_type(SEXP rn);
+R_len_t rownames_size(SEXP rn);
 
 
 #endif